Move onScroll handler from MessageList to Chat
