+++ src/message/MessageList.js
@@ -1,11 +1,8 @@
 import React from 'react';
 import { StyleSheet } from 'react-native';
-import TaggedView from '../native/TaggedView';
 
-import { filterUnreadMessageIds } from '../utils/unread';
-import { registerAppActivity } from '../utils/activity';
+import TaggedView from '../native/TaggedView';
 import { LoadingIndicator } from '../common';
-import { queueMarkAsRead } from '../api';
 import InfiniteScrollView from './InfiniteScrollView';
 import renderMessages from './renderMessages';
 
@@ -18,29 +15,23 @@ const styles = StyleSheet.create({
 export default class MessageList extends React.PureComponent {
   autoScrollToBottom = false;
 
+  static defaultProps = {
+    onScroll: () => {},
+  };
+
   componentWillReceiveProps(nextProps) {
     this.autoScrollToBottom = this.props.caughtUp.newer && nextProps.caughtUp.newer;
   }
 
-  onScroll = e => {
-    if (!e.visibleIds) {
-      return; // temporary fix for Android
-    }
-
-    const { auth, flags, markMessagesRead } = this.props;
-    const visibleMessageIds = e.visibleIds.map(x => +x);
-    const unreadMessageIds = filterUnreadMessageIds(visibleMessageIds, flags);
-
-    if (markMessagesRead && unreadMessageIds.length > 0) {
-      markMessagesRead(unreadMessageIds);
-      queueMarkAsRead(auth, unreadMessageIds);
-    }
-
-    registerAppActivity(auth);
-  };
-
   render() {
-    const { caughtUp, fetching, fetchOlder, fetchNewer, singleFetchProgress } = this.props;
+    const {
+      caughtUp,
+      fetching,
+      fetchOlder,
+      fetchNewer,
+      singleFetchProgress,
+      onScroll,
+    } = this.props;
     const messageList = renderMessages(this.props);
 
     // `headerIndices` tell the scroll view which components are headers
@@ -72,7 +63,7 @@ export default class MessageList extends React.PureComponent {
         onStartReached={fetchOlder}
         onEndReached={fetchNewer}
         autoScrollToBottom={this.autoScrollToBottom}
-        onScroll={this.onScroll}
+        onScroll={onScroll}
       >
         <LoadingIndicator active={fetching.older} caughtUp={caughtUp.older} />
         {messageList}
